lambda1/requirements.txt
boto3
pandas
requests
python-dotenv